perf: register pool, reference-match dispatch, range step inlining; remove spawn/threading
Core VM optimizations:
- Reference-match dispatch: no instruction clone per iteration (InstrPtr)
- Register pool (REG_POOL): reuse Vec allocations across function calls
- Per-frame Vec<Value> registers: owned, no Arc sharing, no AtomicU64
- GC frame scanning via thread-local (CURRENT_FRAMES), remove per-call task_roots
- Range step inlining: reorder instructions so step value loads before Range,
  allowing for-range bypass of heap allocation for step() ranges too

Remove spawn + threading:
- Remove Spawn instruction, SpawnData, parser fields (is_in_spawn etc.)
- Strip RwLock/AtomicU64 from ManagedObject::List/Object (now Vec<Value>)
- Remove parking_lot::RwLock from heap objects (data directly on inner Vec)
- Remove rayon parallel iterators from GC (sequential only)
- Remove JoinSet parameter from execute_bytecode
- Per-frame registers: CallFrame owns Vec<Value>, ReturnTarget stores dst only
- Remove active_registers, TaskRegisters (GC scans frames via thread-local)

Parser bugfixes:
- String pool: value_as_string/with_str fallthrough for pool strings
- else with newlines (skip_newlines before else parsing)
- .step() parser optimization: reorder instructions correctly
- For-range: bypass RangeInfo for direct ranges (no heap alloc)

Examples:
- Remove spawn usage from test files
- Update to sequential execution

+++ ys-cli/src/error_display.rs
@@ -4,12 +4,12 @@ use ys_core::error::JitError;
 /// Pretty-print a YatsuScript error with syntax highlighting logic.
 pub fn display_error(err: &JitError, source: &str) {
     let (msg, line, col) = match err {
-        JitError::Lexing { err, loc } => (err.to_string(), loc.line as usize, loc.col as usize),
-        JitError::Parsing { msg, loc } => (msg.clone(), loc.line as usize, loc.col as usize),
-        JitError::Runtime { msg, loc } => (msg.clone(), loc.line as usize, loc.col as usize),
-        JitError::UnknownVariable { msg, loc } => (msg.clone(), loc.line as usize, loc.col as usize),
+        JitError::Lexing { err, loc } => (err.to_string(), loc.line, loc.col),
+        JitError::Parsing { msg, loc } => (msg.clone(), loc.line, loc.col),
+        JitError::Runtime { msg, loc } => (msg.clone(), loc.line, loc.col),
+        JitError::UnknownVariable { msg, loc } => (msg.clone(), loc.line, loc.col),
         JitError::RedefinitionOfImmutableVariable { msg, loc, orig_line } => {
-            (format!("{} (already defined on line {})", msg, orig_line), loc.line as usize, loc.col as usize)
+            (format!("{} (already defined on line {})", msg, orig_line), loc.line, loc.col)
         }
     };
